+++ __tests__/acceptance/todomvc.test.ts
@@ -32,37 +32,4 @@ describe('Todomvc', () => {
     await elements('#todo-list li').by(have.cssClass('completed')).should(have.texts('a', 'b', 'd'));
     await elements('#todo-list li').by(have.no.cssClass('completed')).should(have.texts('c'));
   });
-
-  it('should not share todos among different users (simulating by different browser contexts)', async () => {
-    await goto('http://todomvc.com/examples/emberjs');
-    await element('#new-todo').type('a').then(perform.press('Enter'));
-    await elements('#todo-list li').should(have.size(1));
-
-    const another = await director.newContext();
-    await another.page.goto('http://todomvc.com/examples/emberjs')
-
-    await another.elements('#todo-list li').should(have.size(0));
-  });
-
-  it('should not share todos among different browser', async () => {
-    await goto('http://todomvc.com/examples/emberjs');
-    await element('#new-todo').type('a').then(perform.press('Enter'));
-    await elements('#todo-list li').should(have.size(1));
-
-    const another = await director.newBrowser();
-    await another.page.goto('http://todomvc.com/examples/emberjs')
-
-    await another.elements('#todo-list li').should(have.size(0));
-  });
-
-  it('should share todos among different tabs', async () => {
-    await goto('http://todomvc.com/examples/emberjs');
-    await element('#new-todo').type('a').then(perform.press('Enter'));
-    await elements('#todo-list li').should(have.texts('a'));
-
-    const another = await director.newPage();
-    await another.page.goto('http://todomvc.com/examples/emberjs')
-
-    await elements('#todo-list li').should(have.texts('a'));
-  });
 });